+++ web/app/themes/lark-child/resources/views/front-page.blade.php
@@ -1,5 +1,7 @@
 @extends('layouts.app')
 
 @section('content')
+
+  {{dd(headingSize(['size' => 'h1', 'mobile.size' => 'something']))}}
   @include('layouts.flexible')
 @endsection