+++ CHANGELOG.md
@@ -313,6 +313,12 @@ All notable changes to `starter-wordpress` will be documented in this file
 - Added `fullAddress()` helper method
 - Added `headingSize()` helper method
 
+## 3.0.26 - 20-04-2021
+
+- Upon switching themes, create 30 default posts to test the blog index styling with
+- Upon switching themes, auto activate plugins defined in the theme configuration file
+- Added ability to set dynamic heading values for both mobile and desktop where necessary using the `headingSize()` helper method
+
 ## To do
 - Create sample WooCommerce, SASS and blade code
 - Add more customisation!